🔨🤖 move the svg tester's helpers out of baker/
The tester built every chart through initGrapherForSvgExport and named
its reference files through buildSvgOutFilename and
grapherSlugToExportFileKey, all exported from baker/ but called by nothing
else. They now live in devTools/svgTester/utils.ts, with the file-key test
moved alongside. dump-data.ts owns its published-graphers query, returning
one map by id since neither caller needed the slug map, and the two unused
exports getGraphersAndRedirectsBySlug and buildSvgOutFilepath are gone.
GrapherImageBaker.tsx keeps grapherToSVG and grapherToPng, the two
functions the admin test page and archival baker call.

The tester no longer imports baker/, so its tsconfig references db
directly. A CI path filter for the svg tester can now watch the grapher
packages and devTools/svgTester/ alone.
